[issue #2] New rule every day on launch.
Now has app settings which are being used to store last viewed rule and
last day that rule was viewed on.

Rule view updates on launch. App's current rule updates on return from
inactive if day is different, but the view is not updating based on this
change yet (without a relaunch of the app).